+++ src/core/internal/modes/dialog.ts
@@ -5,6 +5,7 @@ import * as RpcRequest from 'ox/RpcRequest'
 import * as RpcResponse from 'ox/RpcResponse'
 import * as RpcSchema from 'ox/RpcSchema'
 import { waitForCallsStatus } from 'viem/actions'
+import * as z from 'zod/mini'
 import type { ThemeFragment } from '../../../theme/Theme.js'
 import * as Account from '../../../viem/Account.js'
 import * as Key from '../../../viem/Key.js'
@@ -16,8 +17,6 @@ import * as Mode from '../mode.js'
 import * as Permissions from '../permissions.js'
 import * as PermissionsRequest from '../permissionsRequest.js'
 import type * as Porto from '../porto.js'
-import * as PreCalls from '../preCalls.js'
-import * as Schema from '../schema/schema.js'
 import type * as Token from '../schema/token.js'
 import * as Siwe from '../siwe.js'
 import * as U from '../utils.js'
@@ -34,8 +33,6 @@ export function dialog(parameters: dialog.Parameters = {}) {
 
   const listeners = new Set<(requestQueue: readonly QueuedRequest[]) => void>()
   const requestStore = RpcRequest.createStore()
-
-  let secure: { frame: boolean } | undefined
 
   // Function to instantiate a provider for the dialog. This
   // will be used to queue up requests for the dialog and
@@ -166,7 +163,8 @@ export function dialog(parameters: dialog.Parameters = {}) {
 
             // Convert the key into a permission.
             const permissionsRequest = key
-              ? Schema.encodeSync(PermissionsRequest.Schema)(
+              ? z.encode(
+                  PermissionsRequest.Schema,
                   PermissionsRequest.fromKey(key),
                 )
               : undefined
@@ -204,7 +202,7 @@ export function dialog(parameters: dialog.Parameters = {}) {
               ?.map((permission) => {
                 try {
                   const key_permission = Permissions.toKey(
-                    Schema.decodeSync(Permissions.Schema)(permission),
+                    z.decode(Permissions.Schema, permission),
                   )
                   if (key_permission.id === key?.id)
                     return {
@@ -218,13 +216,6 @@ export function dialog(parameters: dialog.Parameters = {}) {
                 }
               })
               .filter(Boolean) as readonly Key.Key[]
-
-            const { preCalls } = account.capabilities ?? {}
-            if (preCalls)
-              await PreCalls.add(preCalls as PreCalls.PreCalls, {
-                address: account.address,
-                storage,
-              })
 
             const signInWithEthereum_response = await (async () => {
               if (!account.capabilities?.signInWithEthereum) return
@@ -315,9 +306,7 @@ export function dialog(parameters: dialog.Parameters = {}) {
 
         const provider = getProvider(store)
         const result = await provider.request(request)
-        return Schema.decodeSync(RpcSchema_porto.wallet_getAssets.Response)(
-          result,
-        )
+        return z.decode(RpcSchema_porto.wallet_getAssets.Response, result)
       },
 
       async getCallsStatus(parameters) {
@@ -364,16 +353,14 @@ export function dialog(parameters: dialog.Parameters = {}) {
           const result = await provider.request({
             method: 'wallet_getKeys',
             params: [
-              Schema.encodeSync(RpcSchema_porto.wallet_getKeys.Parameters)({
+              z.encode(RpcSchema_porto.wallet_getKeys.Parameters, {
                 address: account.address,
                 chainIds,
               }),
             ],
           })
 
-          return Schema.decodeSync(RpcSchema_porto.wallet_getKeys.Response)(
-            result,
-          )
+          return z.decode(RpcSchema_porto.wallet_getKeys.Response, result)
         })()
 
         return U.uniqBy(
@@ -417,13 +404,8 @@ export function dialog(parameters: dialog.Parameters = {}) {
       },
 
       async grantPermissions(parameters) {
-        const { account, internal } = parameters
-        const {
-          client,
-          config: { storage },
-          request,
-          store,
-        } = internal
+        const { internal } = parameters
+        const { client, request, store } = internal
 
         if (request.method !== 'wallet_grantPermissions')
           throw new Error(
@@ -438,23 +420,17 @@ export function dialog(parameters: dialog.Parameters = {}) {
         })
         if (!key) throw new Error('no key found.')
 
-        const permissionsRequest = Schema.encodeSync(PermissionsRequest.Schema)(
+        const permissionsRequest = z.encode(
+          PermissionsRequest.Schema,
           PermissionsRequest.fromKey(key),
         )
 
         // Send a request off to the dialog to grant the permissions.
         const provider = getProvider(store)
-        const response = await provider.request({
+        await provider.request({
           method: 'wallet_grantPermissions',
           params: [permissionsRequest],
         })
-
-        const { preCalls } = response.capabilities ?? {}
-        if (preCalls)
-          await PreCalls.add(preCalls as PreCalls.PreCalls, {
-            address: account.address,
-            storage,
-          })
 
         return { key }
       },
@@ -466,10 +442,11 @@ export function dialog(parameters: dialog.Parameters = {}) {
 
         const provider = getProvider(store)
 
-        const request =
-          internal.request as typeof RpcSchema_porto.wallet_connect.Request.Encoded & {
-            _decoded: RpcSchema_porto.wallet_connect.Request
-          }
+        const request = internal.request as z.input<
+          typeof RpcSchema_porto.wallet_connect.Request
+        > & {
+          _decoded: RpcSchema_porto.wallet_connect.Request
+        }
 
         if (
           request.method !== 'wallet_connect' &&
@@ -499,7 +476,8 @@ export function dialog(parameters: dialog.Parameters = {}) {
 
           // Convert the key into a permissions request.
           const permissionsRequest = key
-            ? Schema.encodeSync(PermissionsRequest.Schema)(
+            ? z.encode(
+                PermissionsRequest.Schema,
                 PermissionsRequest.fromKey(key),
               )
             : undefined
@@ -525,17 +503,6 @@ export function dialog(parameters: dialog.Parameters = {}) {
             ],
           })
 
-          await Promise.all(
-            accounts.map(async (account) => {
-              const { preCalls } = account.capabilities ?? {}
-              if (!preCalls) return
-              await PreCalls.add(preCalls as PreCalls.PreCalls, {
-                address: account.address,
-                storage,
-              })
-            }),
-          )
-
           return Promise.all(
             accounts.map(async (account) => {
               const adminKeys = account.capabilities?.admins
@@ -545,7 +512,7 @@ export function dialog(parameters: dialog.Parameters = {}) {
                 ?.map((permission) => {
                   try {
                     const key_permission = Permissions.toKey(
-                      Schema.decodeSync(Permissions.Schema)(permission),
+                      z.decode(Permissions.Schema, permission),
                     )
                     if (key_permission.id === key?.id)
                       return {
@@ -602,11 +569,7 @@ export function dialog(parameters: dialog.Parameters = {}) {
 
       async prepareCalls(parameters) {
         const { account, internal } = parameters
-        const {
-          config: { storage },
-          store,
-          request,
-        } = internal
+        const { store, request } = internal
 
         if (request.method !== 'wallet_prepareCalls')
           throw new Error('Cannot prepare calls for method: ' + request.method)
@@ -616,15 +579,9 @@ export function dialog(parameters: dialog.Parameters = {}) {
 
         const feeToken = await resolveFeeToken(internal, parameters)
 
-        const preCalls = await PreCalls.get({
-          address: account.address,
-          storage,
-        })
-
         const provider = getProvider(store)
-        const result = Schema.decodeSync(
+        const result = z.decode(
           RpcSchema_porto.wallet_prepareCalls.Response,
-        )(
           await provider.request({
             ...request,
             params: [
@@ -633,7 +590,6 @@ export function dialog(parameters: dialog.Parameters = {}) {
                 capabilities: {
                   ...request.params?.[0]?.capabilities,
                   feeToken,
-                  preCalls,
                 },
               },
             ],
@@ -675,9 +631,7 @@ export function dialog(parameters: dialog.Parameters = {}) {
 
         // Convert the key into a permission.
         const permissionsRequest = key
-          ? Schema.encodeSync(PermissionsRequest.Schema)(
-              PermissionsRequest.fromKey(key),
-            )
+          ? z.encode(PermissionsRequest.Schema, PermissionsRequest.fromKey(key))
           : undefined
 
         // Send a request off to the dialog to prepare the upgrade.
@@ -773,23 +727,11 @@ export function dialog(parameters: dialog.Parameters = {}) {
           merchantUrl,
           requiredFunds,
         } = parameters
-        const {
-          config: { storage },
-          client,
-          store,
-          request,
-        } = internal
+        const { client, store, request } = internal
 
         const provider = getProvider(store)
 
         const feeToken = await resolveFeeToken(internal, parameters)
-
-        const preCalls = secure?.frame
-          ? await PreCalls.get({
-              address: account.address,
-              storage,
-            })
-          : undefined
 
         // Try and extract an authorized key to sign the calls with.
         const key = await Mode.getAuthorizedExecuteKey({
@@ -808,7 +750,7 @@ export function dialog(parameters: dialog.Parameters = {}) {
           try {
             // TODO: use eventual Viem Action.
             const req = await provider.request(
-              Schema.encodeSync(RpcSchema_porto.wallet_prepareCalls.Request)({
+              z.encode(RpcSchema_porto.wallet_prepareCalls.Request, {
                 method: 'wallet_prepareCalls',
                 params: [
                   {
@@ -819,7 +761,6 @@ export function dialog(parameters: dialog.Parameters = {}) {
                         : undefined),
                       feeToken,
                       merchantUrl,
-                      preCalls,
                       requiredFunds,
                     },
                     chainId: client.chain.id,
@@ -854,11 +795,6 @@ export function dialog(parameters: dialog.Parameters = {}) {
                   signature,
                 },
               ],
-            })
-
-            await PreCalls.clear({
-              address: account.address,
-              storage,
             })
 
             const response = result[0]
@@ -902,15 +838,9 @@ export function dialog(parameters: dialog.Parameters = {}) {
                 capabilities: {
                   feeToken,
                   merchantUrl,
-                  preCalls,
                 },
               },
             ],
-          })
-
-          await PreCalls.clear({
-            address: account.address,
-            storage,
           })
 
           return { id }
@@ -927,15 +857,9 @@ export function dialog(parameters: dialog.Parameters = {}) {
                   ...request.params?.[0]?.capabilities,
                   feeToken,
                   merchantUrl,
-                  preCalls,
                 },
               },
             ],
-          })
-
-          await PreCalls.clear({
-            address: account.address,
-            storage,
           })
 
           return result
@@ -945,12 +869,8 @@ export function dialog(parameters: dialog.Parameters = {}) {
       },
 
       async sendPreparedCalls(parameters) {
-        const { account, internal } = parameters
-        const {
-          config: { storage },
-          store,
-          request,
-        } = internal
+        const { internal } = parameters
+        const { store, request } = internal
 
         if (request.method !== 'wallet_sendPreparedCalls')
           throw new Error(
@@ -962,12 +882,6 @@ export function dialog(parameters: dialog.Parameters = {}) {
 
         const provider = getProvider(store)
         const result = await provider.request(request)
-
-        if (account.address)
-          await PreCalls.clear({
-            address: account.address,
-            storage,
-          })
 
         const id = result[0]?.id
         if (!id) throw new Error('id not found')
@@ -1051,10 +965,6 @@ export function dialog(parameters: dialog.Parameters = {}) {
         internal,
         theme,
         themeController,
-      })
-
-      dialog.secure().then((s) => {
-        secure = s
       })
 
       const unsubscribe = store.subscribe(