+++ src/core/internal/provider.test.ts
@@ -17,6 +17,7 @@ import {
   setCode,
   signTypedData,
   verifyHash,
+  setBalance as viem_setBalance,
   waitForCallsStatus,
   waitForTransactionReceipt,
 } from 'viem/actions'
@@ -1873,6 +1874,10 @@ describe.each([['relay', Mode.relay]] as const)('%s', (type, mode) => {
         })
         const address = account!.address
 
+        await viem_setBalance(client, {
+          address,
+          value: Value.fromEther('10000'),
+        })
         await setBalance(client, {
           address,
           value: Value.fromEther('10000'),
@@ -2160,6 +2165,7 @@ describe.each([['relay', Mode.relay]] as const)('%s', (type, mode) => {
                   to: contracts.exp1.address,
                 },
               ],
+              chainId: Anvil.enabled ? undefined : '0x2105',
               from: address,
               version: '1',
             },
@@ -3476,7 +3482,7 @@ describe.each([['relay', Mode.relay]] as const)('%s', (type, mode) => {
         ],
       })
 
-      const signature = await signTypedData(walletClient, typedData)
+      const signature = await signTypedData(walletClient, typedData as never)
 
       const result = await porto.provider.request({
         method: 'wallet_sendPreparedCalls',
@@ -3521,6 +3527,21 @@ describe.each([['relay', Mode.relay]] as const)('%s', (type, mode) => {
     ).rejects.toThrowError(
       'The provider does not support the requested method.',
     )
+  })
+
+  test('behavior: invalid params', async () => {
+    const porto = getPorto()
+    await porto.provider.request({
+      method: 'wallet_connect',
+      params: [{ capabilities: { createAccount: true } }],
+    })
+    await expect(() =>
+      porto.provider.request({
+        method: 'eth_sendTransaction',
+        // @ts-expect-error
+        params: [{ to: 1 }],
+      }),
+    ).rejects.matchSnapshot()
   })
 })